refactor(cli): isolate init provider metadata
