+++ packages/cli/src/init.ts
@@ -1,5 +1,13 @@
 import { mkdir, writeFile } from 'node:fs/promises'
 import path from 'node:path'
+import {
+  PROVIDER_DEFAULT_MODEL,
+  PROVIDER_ENV_KEY,
+  PROVIDER_IMPORT,
+} from './init-providers'
+import type { Provider } from './init-providers'
+
+export type { Provider } from './init-providers'
 
 export type StarterKind =
   | 'react'
@@ -15,17 +23,6 @@ export type StarterKind =
   | 'expo'
   | 'deno-deploy'
   | 'angular'
-export type Provider =
-  | 'openai'
-  | 'anthropic'
-  | 'gemini'
-  | 'ollama'
-  | 'deepseek'
-  | 'grok'
-  | 'kimi'
-  | 'groq'
-  | 'openrouter'
-  | 'demo'
 export type ToolKind = 'web_search' | 'filesystem' | 'shell'
 export type MemoryKind = 'none' | 'file' | 'sqlite'
 export type PackageManager = 'pnpm' | 'npm' | 'yarn' | 'bun'
@@ -45,44 +42,6 @@ interface RenderContext {
   tools: ToolKind[]
   memory: MemoryKind
   pm: PackageManager
-}
-
-const PROVIDER_IMPORT: Record<Exclude<Provider, 'demo'>, string> = {
-  openai: 'openai',
-  anthropic: 'anthropic',
-  gemini: 'gemini',
-  ollama: 'ollama',
-  deepseek: 'deepseek',
-  grok: 'grok',
-  kimi: 'kimi',
-  groq: 'groq',
-  openrouter: 'openrouter',
-}
-
-const PROVIDER_DEFAULT_MODEL: Record<Provider, string> = {
-  openai: 'gpt-4o-mini',
-  anthropic: 'claude-sonnet-4-6',
-  gemini: 'gemini-2.5-flash',
-  ollama: 'llama3.1',
-  deepseek: 'deepseek-chat',
-  grok: 'grok-4.20-0309-non-reasoning',
-  kimi: 'kimi-k2-0711-preview',
-  groq: 'openai/gpt-oss-120b',
-  openrouter: '~anthropic/claude-haiku-latest',
-  demo: 'demo',
-}
-
-const PROVIDER_ENV_KEY: Record<Provider, string | null> = {
-  openai: 'OPENAI_API_KEY',
-  anthropic: 'ANTHROPIC_API_KEY',
-  gemini: 'GEMINI_API_KEY',
-  ollama: null,
-  deepseek: 'DEEPSEEK_API_KEY',
-  grok: 'XAI_API_KEY',
-  kimi: 'KIMI_API_KEY',
-  groq: 'GROQ_API_KEY',
-  openrouter: 'OPENROUTER_API_KEY',
-  demo: null,
 }
 
 function adapterCall(provider: Provider, prefix = 'process.env'): string {
